Differentiate between the different base URLs for interacting with the Reddit API
